rearrange modules files
